+++ src/tiles/ordinary.h
@@ -10,7 +10,7 @@ class Ordinary : public Tile, public IStepOnTile {
 
 public:
   Ordinary();
-	Ordinary(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint8_t type);
+	Ordinary(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint16_t type);
   ~Ordinary();
 
   void TileSteppedOnEvent() override;
